Hook up save and load buttons
These use a technique where the save/load functionality is passed from the component to its children via a function argument.  I opted for this rather than a higher order component because it is more readable and requires fewer intermediary components.

This could just as easily been done with a higher order component and would be happy to change it to that instead.

There is a performance implication since there is a function being defined in render. However the menu bar is not a highly re-rendered component, so I thought it would be ok in this case.